feat: add DC-DC control loop
A PI control loop has been implemented. vDcDcControl(Vout, VoutTarget) has to be called periodically: the controller adapts DA (duty cycle of timer A) and DB (duty cycle of timer B) based on the error (difference between Vout and VoutTarget), and switches operating mode (buck, mixed, boost) if it is necessary to reach VoutTarget. vDcDcControl() is called (in vDcDcTask()) every 5ms currently.

The following functions can be used to set operating mode and duty cycle manually if necessary (they have to be called only once):
- vSetDutyCycleBuck(DA)
- vSetDutyCycleBoost(DB)
- vSetDutyCycleMixed(DA, DB)

vSetAdcTriggerPoint(Val) sets the ADC trigger event. This depends on the operating mode and the duty cycle because the voltages are measured before or after the PWM rising or falling edge, this way the measurement noise can be minimalized.

Examples:
- vDcDcControl(ulVoutPhys, 5555u); - set the output voltage to 5555mV
- vSetDutyCycleBuck(5022u); - set buck operating mode with 50.22% duty cycle
- vSetAdcTriggerPoint(2534u); - set ADC trigger event to 25.34% (ADC is triggered at 25.34% of the PWM period)

Warning:
- extreme Vin/Iin and Vout/Iout values can damage the converter (there is no SW check implemented yet)
- the PI control loop and operating mode switch is working but is not thoroughly tested yet

For more information refer to AN4449 and AN4539.